+++ src/index.php
@@ -1,6 +1,8 @@
 <?php
 session_start();
-require_once("./modules/fileStats.php");
+
+
+
 
 /* -------------------------------------------------------------------------- */
 /*                              SESSION VARIABLES                             */
@@ -19,40 +21,9 @@ if (!isset($_SESSION["searchFiles"])) {
     $_SESSION["searchFiles"] = array();
 }
 
-
-
+require_once("./modules/searchFile.php");
 // unset($_SESSION);
 // session_destroy();
-
-
-$currPath = $_SESSION["currentPath"];
-
-$_SESSION["searchFiles"] = array();
-searchFilesInside($currPath);
-
-
-
-function searchFilesInside($path_file)
-{
-    if (is_dir($path_file)) {
-        $files = array_diff(scandir($path_file), [".", "..", ".DS_Store"]);
-        foreach ($files as $file) {
-
-            searchFilesInside("$path_file/$file");
-        }
-        $pathArray = explode("/", $path_file);
-        $fileName = end($pathArray);
-        $fileArray = getFileStats($path_file, $fileName);
-        array_push($_SESSION["searchFiles"], $fileArray);
-    } else {
-        $pathArray = explode("/", $path_file);
-        $fileName = end($pathArray);
-        $fileArray = getFileStats($path_file, $fileName);
-        array_push($_SESSION["searchFiles"], $fileArray);
-    }
-}
-// echo "<pre>" . print_r($_SESSION["searchFiles"], true) . "</pre>";
-
 
 ?>
